[fix] 코드리뷰; ResultDto 통일, 파일명 수정
