--- packages/dictionary-api/src/modules/word/dtos/create-word-dto.ts
@@ -1,19 +1,0 @@
-import { OmitType } from '@nestjs/swagger';
-import { Type } from 'class-transformer';
-import { ValidateNested } from 'class-validator';
-import { ResultDto } from 'src/common/common.dto';
-import { IsArrayOfObjects } from 'src/decorators/nested-dto.decorator';
-import { Word } from 'src/schemas/word.schema';
-import { Tag, WordDto } from './word-dto';
-
-export class InputCreateWordDto extends WordDto {
-  @IsArrayOfObjects()
-  @ValidateNested()
-  @Type(() => Tag)
-  tags: Tag[];
-}
-
-export class OutputCreateWordDto implements ResultDto {
-  items: Word;
-  message: string;
-}
