+++ packages/dictionary-api/src/modules/word/dtos/create-word.dto.ts
@@ -1,0 +1,3 @@
+import { WordDto } from './word.dto';
+
+export class InputCreateWordDto extends WordDto {}
