+++ packages/dictionary-api/src/modules/word/dtos/get-word.dto.ts
@@ -1,0 +1,4 @@
+import { PickType } from '@nestjs/swagger';
+import { WordDto } from './word.dto';
+
+export class InputGetWordDto extends PickType(WordDto, ['title'] as const) {}
